+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## **Unreleased**
+**Hearthstone**:
+- Fixed King Llane appearing in the wrong player's decklist.
+
 ## **Release v1.51.1 - 2026-03-26**
 **Hearthstone**:
 - New Premium feature: Mulligan G-V2!\