+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## **Unreleased**
 **Hearthstone**:
+- Druid Imbue Counter now shows spell progress.
 - Fixed King Llane appearing in the wrong player's decklist.
 
 ## **Release v1.51.1 - 2026-03-26**